requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
Django
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
email_validator==2.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rich==13.7.1
rpds-py==0.20.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.35.1
swagger-spec-validator==3.0.4
typer==0.12.3
typing_extensions==4.12.2
uritemplate==4.1.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
ipython